Add the concept of status to the dispatcher
Fix Issue #7

The dispatcher now understand the concept of run status and will act
differently based on the current status of the run. This will prevent a
dispatcher from repeatedly publish jobs to the queue after it crashes
and restarted by the Kubernetes.

Additionally in this commit, major refactory is done to the dispatcher
to move logics closer to the data models.